LNX-341: Add support ASN and subnets to the mirror service
- Implementation of hyper clouds part
