+++ src/backend/db/dbmigrations/versions/3df99e9174f4_cloud_subnets.py
@@ -1,0 +1,32 @@
+"""cloud_subnets
+
+Revision ID: 3df99e9174f4
+Revises: 819b9d9d7121
+Create Date: 2021-07-16 09:22:46.057750
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = '3df99e9174f4'
+down_revision = '819b9d9d7121'
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column('mirrors', sa.Column('cloud_type', sa.String(),
+                                       nullable=True))
+    op.add_column('mirrors', sa.Column('cloud_region', sa.String(),
+                                       nullable=True))
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_column('mirrors', 'cloud_region')
+    op.drop_column('mirrors', 'cloud_type')
+    # ### end Alembic commands ###
